requirements.txt
Flask
psycopg2-binary
gunicorn
Werkzeug
python-dotenv
